Add functions for data visualization and summary (#4)
Adds functions for creating a visdat plot and a summary table. The summary table is a mix of skimr and a custom function. The skimr package was used to create an initial table summarizing the variables in the file. The custom function gathers the values that appear for a variable and the number of times they appear into a string. This is then attached to the basic skimr table.