Test get_all_file_templates() and update get_template() (#14)
* Update get_template to return NA for metadataType is NA

There may be times where the documentation rows haven't been filtered out of the fileview that's passed to get_all_file_templates(). In this case, the metadataType will be NA and there would be a problem with trying to assign NULL to the column. Instead, just pass back NA. In terms of fileviews, there should not be a NULL metadataType that is passed in, only NA or the exact type.

* Test get_all_file_templates()

Add tests for get_all_file_templates(). Not sure how to test get_all_file_data() yet since relies on get_data() which uses synapser to download a file.