Add annotations link to config to function with dccvalidator (#44)
Annotations link used in behavior for some of the checks in check_all(). In dccvalidator, this is for the user to see what annotations are accepted. In dccmonitor, this is less necessary since the data curators are the only ones who would see the link, but dccmonitor would crash without a link to refer to in the config. While it could have been any text, I put the actual links.